Updates made for newer API versions (> v 11.1)
loginURL and portalURL updated (you will need to insert your own portalURL as assigned by Salesforce).  Also results are returned with a slight difference in syntax by newer versions of the API, so  made some corrections in parseQueryXML to accommodate. Finally, newer API doesn't return field names (column names once converted to query) when null for some objects, so added some logic to check for the existence of field names when populating the query objects and to add column(s) if not present.